Fix terminal lifecycle issues. (#734)
This could cause issues if XTerm was remounted or when the user changed
language.

Closes https://github.com/microbit-foundation/python-editor-next/issues/677
